Initial POC commit
This implements the initial POC version of pg_pgbouncer. It does the
basic job of spawning PgBouncer processes from inside Postgres, but it's
definitely not production ready. Mainly because there's a bunch of
issues with the current code quality and feature-set. A major issue is
that there is no automated testing, except for checknig that the code
compiles. I'll try to list my other intended improvements in GitHub
issues.